LayoutTests/accessibility/mac/math-alttext-expected.txt

Fe
1
x
Co
x
Sb
2
This tests that a math element will be exposed like an ARIA math element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS math.role is 'AXRole: AXGroup'
PASS math.subrole is 'AXSubrole: AXDocumentMath'
PASS math.description is 'AXDescription: F e sub 1  minus x, end subscript. C o sub x, S b sub 2'
PASS successfullyParsed is true

TEST COMPLETE
